+++ sctoolbox/utils/assemblers.py
@@ -17,7 +17,6 @@ import sctoolbox.utils as utils
 #####################################################################
 
 def assemble_from_h5ad(h5ad_files,
-                       qc_columns,
                        merge_column='sample',
                        coordinate_cols=None,
                        set_index=True,
@@ -65,20 +64,6 @@ def assemble_from_h5ad(h5ad_files,
                                 "coordinate_cols")  # Check that coordinate_cols are in adata.var)
 
         utils.format_adata_var(adata, coordinate_cols, coordinate_cols)
-
-        # Add information to the infoprocess
-        # cr.build_infor(adata, "Input_for_assembling", h5ad_path)
-        # cr.build_infor(adata, "Strategy", "Read from h5ad")
-
-        print('add existing adata.obs columns to infoprocess:')
-        print()
-        for key, value in qc_columns.items():
-            if value is not None:
-                print(key + ':' + value)
-                if value in adata.obs.columns:
-                    utils.build_legend(adata, key, value)
-                else:
-                    print('column:  ' + value + ' is not in adata.obs')
 
         # check if the barcode is the index otherwise set it
         utils.barcode_index(adata)